Citations shanti sources bug/manu 5001/drc2r (#4)
* Fix support for SHANTI sources

There were some inconsistencies when using the model sources with the
asset index where the Source Resource was being taken as if it belongs
to a database. and fixes the code in the bibliographical_reference
method parsing the Solr service URL.

* Improves Sources to support ActiveRecord Association

Added module ActiveRecordRelatable that allows sources to be associated
in a polomorphic association.

* Fix Source uid prefix

* Fix Source uid prefix

* Fix source production shanti source service name

* Updates engine version

The branch is going to be merged to master so in preparation we are
going to increase the engine version.