Mandala rails theme sync/manu 4933/drc2r (#3)
* CSS sync with the Mandala sites

The css has been updated to use the css from mandala apps.

* Fix bug with chop top of header

Fix a bug that occured when the content of the page was smaller than the
sidebar, when that occurred the top of the page got chopped off, now its
fixed.

* Fix the code for the fly-out resizing

Did a code cleanup and fixed the values to support resizing with the
jquery plugin Resize.

* Corrected file names for JavaScript form Mandala

Correctly named the main.js and search.js to their equivalent in Mandala
so it is easier to map them. And removed some old copies of main that
were no longer used.

* Citations shanti sources bug/manu 5001/drc2r (#4)

* Fix support for SHANTI sources

There were some inconsistencies when using the model sources with the
asset index where the Source Resource was being taken as if it belongs
to a database. and fixes the code in the bibliographical_reference
method parsing the Solr service URL.

* Improves Sources to support ActiveRecord Association

Added module ActiveRecordRelatable that allows sources to be associated
in a polomorphic association.

* Fix Source uid prefix

* Fix Source uid prefix

* Fix source production shanti source service name

* Updates engine version

The branch is going to be merged to master so in preparation we are
going to increase the engine version.

* CSS sync with the Mandala sites

The css has been updated to use the css from mandala apps.

* Fix bug with chop top of header

Fix a bug that occured when the content of the page was smaller than the
sidebar, when that occurred the top of the page got chopped off, now its
fixed.

* Fix the code for the fly-out resizing

Did a code cleanup and fixed the values to support resizing with the
jquery plugin Resize.

* Corrected file names for JavaScript form Mandala

Correctly named the main.js and search.js to their equivalent in Mandala
so it is easier to map them. And removed some old copies of main that
were no longer used.

* CSS sync with the Mandala sites

The css has been updated to use the css from mandala apps.

* Fix bug with chop top of header

Fix a bug that occured when the content of the page was smaller than the
sidebar, when that occurred the top of the page got chopped off, now its
fixed.

* Fix the code for the fly-out resizing

Did a code cleanup and fixed the values to support resizing with the
jquery plugin Resize.

* Corrected file names for JavaScript form Mandala

Correctly named the main.js and search.js to their equivalent in Mandala
so it is easier to map them. And removed some old copies of main that
were no longer used.

* Update engine version in preparation to merge back to master