Add logic to top menu for disabling main page clicks when open (#5)
There is a bug when clicking on the menu and the menu is over any other
item that has a click action, the click is propagating to them. Build
the logic to add a class disabled-by-top-menu-open to specif items that
were causing problem. This only prevents the clicks going through when
the first level of the menu is open, the event clicks should stop
propagation on their own code, this was done on kmaps_engine for the
public view list partial.